% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{bandicoot.glfbm}
\alias{bandicoot.glfbm}
\title{A dartR object created via the read.dart functions}
\format{
dartRt object
}
\usage{
bandicoot.glfbm
}
\description{
This a test data set to test the validity of functions within dartR and is
based on a DArT SNP data set of simulated bandicoots across Australia. It
contains 96 individuals and 1000 SNPs. this version has a fbm matrix
}
\author{
Bernd Gruber (bugs? Post to
\url{https://groups.google.com/d/forum/dartr}
}
\keyword{datasets}
