% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{platy}
\alias{platy}
\title{Example data set as text file to be imported into a genlight object}
\format{
csv
}
\description{
Check ?read.genetable in pacakge PopGenReport for details on the format.
}
\author{
Bernd Gruber (bugs? Post to
\url{https://groups.google.com/d/forum/dartr}
}
\keyword{datasets}
