% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{possums.glfbm}
\alias{possums.glfbm}
\title{A simulated dartR object created to run a landscape genetic example}
\format{
dartR object with fbm matrix
}
\usage{
possums.glfbm
}
\description{
This a test data set to run a landscape genetics example. It contains 10
populations of 30 individuals each and each individual has 300 loci. There
are no covariates for individuals or loci. this version has a fbm matrix
}
\author{
Bernd Gruber (bugs? Post to
\url{https://groups.google.com/d/forum/dartr}
}
\keyword{datasets}
