% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{testset2.gl}
\alias{testset2.gl}
\title{A genlight object with a known-pedigree captive population, for kinship analyses}
\format{
genlight object
}
\usage{
testset2.gl
}
\description{
An augmented companion to testset.gl. 274 individuals, 755 loci in 31
populations. Extends testset.gl with a captive-bred population
(EmmacCaptBred, n = 24) with known pedigree structure - parent-offspring,
full-sib, half-sib and first-cousin pairs bred from six wild-caught
founders drawn from six different populations - for testing kinship-based
analyses. Pedigree truth (sire, dam, cohort) and COLONY-style
offspring/mother/father flags are recorded in ind.metrics. 500 of the
loci are simulated (CloneID prefix SIM) with intermediate allele
frequencies and Balding-Nichols population structure so that kinship
classes are statistically recoverable; the original 255 DArT loci and
250 wild individuals are unchanged from testset.gl.
}
\author{
Custodian: Arthur Georges (bugs? Post to
\url{https://groups.google.com/d/forum/dartr}
}
\keyword{datasets}
