% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{zzz}
\alias{zzz}
\title{dartR.data
Imports data for examples in dartR.base}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
zzz
}
\description{
dartR.data
Imports data for examples in dartR.base
}
\keyword{datasets}
